Relocated a class
