+++ src/main/java/be/seeseemelk/mockbukkit/command/MockCommandMap.java
@@ -1,7 +1,9 @@
-package be.seeseemelk.mockbukkit;
+package be.seeseemelk.mockbukkit.command;
 
 import org.bukkit.command.CommandMap;
 import org.bukkit.command.SimpleCommandMap;
+
+import be.seeseemelk.mockbukkit.ServerMock;
 
 public class MockCommandMap extends SimpleCommandMap implements CommandMap
 {
